+++ changelog.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Changed 
+### Changed 
 - [Improve inout parameter which supports array types for call procedure](https://github.com/ballerina-platform/ballerina-standard-library/issues/1516)
 - [Add completion type as nil in SQL query return stream type](https://github.com/ballerina-platform/ballerina-standard-library/issues/1654)
+
+### Added
+- [Add support for queryRow](https://github.com/ballerina-platform/ballerina-standard-library/issues/1604)
 
 [0.6.0-beta.2] - 2021-06-02