+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [Add support for queryRow](https://github.com/ballerina-platform/ballerina-standard-library/issues/1604)
+- [Remove support for string parameter in APIs](https://github.com/ballerina-platform/ballerina-standard-library/issues/2010)
 
 [0.6.0-beta.2] - 2021-06-02